+++ app/Http/Controllers/web/adminlte/UserController.php
@@ -9,9 +9,9 @@ use Illuminate\Support\Facades\Hash;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\PermissionRegistrar;
 use Spatie\Permission\Traits\HasRoles;
-
 class UserController extends \App\Http\Controllers\Controller
 {
+    use HasRoles;
     /**
      * Display a listing of the resource.
      *
@@ -22,10 +22,10 @@ class UserController extends \App\Http\Controllers\Controller
         //here we can see all the users
         // you need to learn more about auth and middleware
 
-       // if(auth()->user()->hasAnyRole(['admin'])){
+       if(auth()->user()->hasAnyRole(['admin'])){
         $users= user::paginate(10);
         return view('user\users',['users'=>$users]);
-   // }else{abort(403);}
+   }else{abort(403);}
     }
     /**
      * Show the form for creating a new resource.
@@ -34,12 +34,11 @@ class UserController extends \App\Http\Controllers\Controller
      */
     public function create()
     {
-       // if(auth()->user()->hasAnyRole(['admin'])){ 
-     
-            // $role= Role::all();
-            return view('user\create');//['role'=>$role]);
+       if(auth()->user()->hasAnyRole(['admin'])){ 
+            $role= Role::all();
+            return view('user\create',['role'=>$role]);
             
-      //  }else{abort(403);}
+       }else{abort(403);}
     }
 
     /**
@@ -52,9 +51,21 @@ class UserController extends \App\Http\Controllers\Controller
       {
           //here we can save the new data
 
-      //    if(auth()->user()->hasAnyRole(['admin'])){ 
+         if(auth()->user()->hasAnyRole(['admin'])){ 
+        $validator=validator()->make($request->all(),[
+            "name"=>'required',
+            "password"=>'required|min:8',
+            "email"=>'required|unique:users',
+            "role"=>'required',
+        ]
+        ) ;
+        if ($validator->fails()){
+            $role= Role::all();
+            $messages=$validator->errors();
+            return view('user\create',['role'=>$role,'messages'=>$messages]);
+        }
               $user= new user;
-             // $user->assignRole($request->role);
+             $user->assignRole($request->role);
               $user->name=$request->name;
               $user->email=$request->email;
               $user->password=bcrypt($request->password);
@@ -62,7 +73,7 @@ class UserController extends \App\Http\Controllers\Controller
               $users= user::paginate(10);
     
               return redirect(url('/users'));
-      //  }else{abort(403);}
+       }else{abort(403);}
       }
   
       /**
@@ -85,10 +96,10 @@ class UserController extends \App\Http\Controllers\Controller
       public function edit($id)
       {
           //here u can edit ur users
-       //   if(auth()->user()->hasAnyRole(['admin'])){ 
-             // $role= Role::all();
-              return view('user/edit',['id'=>$id]);//,'role'=>$role]);
-      //  }else{abort(403);}
+         if(auth()->user()->hasAnyRole(['admin'])){ 
+             $role= Role::all();
+              return view('user/edit',['id'=>$id,'role'=>$role]);
+       }else{abort(403);}
       }
   
       /**
@@ -100,11 +111,11 @@ class UserController extends \App\Http\Controllers\Controller
        */
       public function update(Request $request, $id)
       {
-      //  if(auth()->user()->hasAnyRole(['admin'])){ 
+       if(auth()->user()->hasAnyRole(['admin'])){ 
             //here we put the new edited name of users
             $user= user::where('id',$id)->first();
-           // if($request->role)
-           // $user->syncRoles($request->role);
+           if($request->role)
+           $user->syncRoles($request->role);
             if($request->name)
             $user->name=$request->name;
             if($request->email)
@@ -119,7 +130,7 @@ class UserController extends \App\Http\Controllers\Controller
           
   
             return redirect(url('/users'));
-      //  }else{abort(403);}
+       }else{abort(403);}
       }
   
       /**
@@ -131,11 +142,11 @@ class UserController extends \App\Http\Controllers\Controller
       public function destroy($id)
       {
           //here we can delete the user we added
-       //   if(auth()->user()->hasAnyRole(['admin'])){ 
+         if(auth()->user()->hasAnyRole(['admin'])){ 
               $user=user::where('id',$id);
               $user->delete();
               $users= user::paginate(10);
               return redirect(url('/users'));
-     //   }else{abort(403);}
+       }else{abort(403);}
       }
   }
